Put items collection into separate users

+++ lib/repositories/items_repository.dart
@@ -1,9 +1,16 @@
 import 'package:cantwait28/features/models/item_model.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 
 class ItemsRepository {
   Stream<List<ItemModel>> getItemsStream() {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('User is not logged in');
+    }
     return FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('items')
         .orderBy('release_date')
         .snapshots()
@@ -21,37 +28,57 @@ class ItemsRepository {
     });
   }
 
-    Future<void> delete({required String id}) {
+  Future<void> delete({required String id}) {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('User is not logged in');
+    }
     return FirebaseFirestore.instance
-          .collection('items')
-          .doc(id)
-          .delete();
+        .collection('users')
+        .doc(userID)
+        .collection('items')
+        .doc(id)
+        .delete();
   }
 
   Future<ItemModel> get({required String id}) async {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('User is not logged in');
+    }
     final doc = await FirebaseFirestore.instance
-          .collection('items')
-          .doc(id)
-          .get();
+        .collection('users')
+        .doc(userID)
+        .collection('items')
+        .doc(id)
+        .get();
     return ItemModel(
-       id: doc.id,
-       title: doc['title'],
-       imageURL: doc['image_url'],
-       releasedate: (doc['release_date'] as Timestamp).toDate(),
+      id: doc.id,
+      title: doc['title'],
+      imageURL: doc['image_url'],
+      releasedate: (doc['release_date'] as Timestamp).toDate(),
     );
   }
 
-    Future<void> add(
+  Future<void> add(
     String title,
     String imageURL,
     DateTime releaseDate,
   ) async {
-      await FirebaseFirestore.instance.collection('items').add(
-        {
-          'title': title,
-          'image_url': imageURL,
-          'release_date': releaseDate,
-        },
-      );
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('User is not logged in');
+    }
+    await FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
+        .collection('items')
+        .add(
+      {
+        'title': title,
+        'image_url': imageURL,
+        'release_date': releaseDate,
+      },
+    );
   }
 }
